+++ src/components.ts
@@ -56,6 +56,7 @@ export * from './sticky';
 export * from './swiper';
 export * from './table';
 export * from './tag';
+export * from './typography';
 export * from './qrcode';
 
 // 反馈