+++ vitest.config.mjs
@@ -12,16 +12,18 @@ export default mergeConfig(
     test: {
       include:
         /**
-         * 快照涵盖 `__test__/*.test.jsx`
+         * 快照涵盖 `__test__/demo.test.jsx`
          * 生成demo测试文件： npm run test:demo
          * 生成快照：npm run test:snap || npm run test:snap-update
          *
-         * 测试用例检测 && GUI && 覆盖率报告，仅涵盖 `__test__/index.test.jsx`
+         * 测试用例检测 && GUI && 覆盖率报告，涵盖 `__test__/` 下除 demo 外的所有测试文件
          * 单测： npm run test:unit
          * 覆盖率报告： npm run test:unit-coverage
          * GUI：npm run test:unit-gui
          */
-        [`src/**/__test__/${testSnap ? 'demo' : 'index'}.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}`],
+        testSnap
+          ? [`src/**/__test__/demo.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}`]
+          : [`src/**/__test__/!(demo).{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}`],
       globals: true,
       environment: 'jsdom',
       testTimeout: 5000,